+++ frontend/src/app/globals.css
@@ -2,27 +2,12 @@
 
 :root {
   --background: #191F30;
-  --foreground: #171717;
-}
-
-@theme inline {
-  --color-background: var(--background);
-  --color-foreground: var(--foreground);
-  --font-sans: var(--font-geist-sans);
-  --font-mono: var(--font-geist-mono);
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --background: #0a0a0a;
-    --foreground: #ededed;
-  }
 }
 
 body {
   background: var(--background);
-  color: var(--foreground);
-  font-family: Inria Sans, Helvetica, sans-serif;
+  font-family: var(--font-inria-sans), Helvetica, sans-serif;
+  padding: 3rem 2rem;
 }
 
 .custom-alert {
@@ -62,4 +47,20 @@ body {
     opacity: 1;
     transform: scale(1);
   }
+}
+
+.btn-yellow {
+  color: #FAFF00;
+  border: 1px solid #FAFF00;
+  padding: 4px 33px;
+  cursor: pointer;
+  border-radius: 8px;
+  background: transparent;
+}
+
+.custom-input {
+  width: 100%;
+  background-color: #D9D9D9;
+  border-radius: 0.5rem;
+  padding: 0.5rem;
 }
