+++ frontend/src/app/layout.tsx
@@ -1,32 +1,22 @@
 import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
+import { Inria_Sans } from "next/font/google";
 import "./globals.css";
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
+const inriaSans = Inria_Sans({
+  weight: ["300"],
+  variable: "--font-inria-sans",
   subsets: ["latin"],
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "Splity",
   description: "Generated by create next app",
 };
 
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+export default function RootLayout({children,}: Readonly<{children: React.ReactNode;}>) {
   return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
+    <html lang="en" className="h-[912px] flex justify-center bg-[url('/bg-retro.jpg')] bg-repeat">
+      <body className={`${inriaSans.variable} antialiased w-[417px] shadow-2xl`}>
         {children}
       </body>
     </html>
