+++ frontend/src/components/FormRegister/validation.ts
